Modelling/Model/requirements.txt
# Core Libraries
numpy==1.26.0
pandas==2.2.3
matplotlib==3.8.0
scikit-learn==1.3.1
tqdm==4.67.0
typing-extensions==4.8.0

# PyTorch
torch==2.1.0
torchvision==0.16.0

# TensorFlow dan Related Libraries
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-addons==0.22.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.23.0
tensorflowjs

# ONNX Related Libraries
onnx==1.17.0
onnx-tf==1.10.0
onnxruntime==1.20.1
tf2onnx==1.15.1
onnx-graphsurgeon==0.5.2
sng4onnx==1.0.4
onnx2tf==1.10.0

# Additional Libraries
setuptools
Flask==2.0.1
Pillow==10.4.0
gunicorn==20.1.0
scipy==1.11.2
Werkzeug==2.0.1
flask_cors
